requirements.txt
groq>=0.12.0
httpx>=0.27.0
python-dotenv>=1.0.0
Pillow>=10.0.0
markdown>=3.5.0
gradio>=4.19.0
